+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## unreleased
 
 - Added methods `known?`, `status_type`, `body`, `headers` to `RuntimeResponse` to inspect the parsed response
+- Add `OpenapiFirst::ParseError` which is raised by low-level interfaces like `request.body` if the body could not be parsed.
 
 ## 1.0.0